requirements.txt
pandas
numpy
pyyaml
pymysql
scipy
implicit
lightfm
annoy
sklearn
scikit-surprise